Updated Professional Description and Tag Support
Added Tags for the header to every page, and added support for that in the python file which manages that.
