+++ W-TheHallowedEve.html
@@ -26,12 +26,12 @@
     
     <!-- Sidebar -->
     
-    <!-- Note: Content of Sidebar copied from 'index.html',#e8e9ea#f8f9fa
+    <!-- Note: Content of Sidebar copied from 'index.html',#f8f9fa
            managed by a python script -->
     
       
-    <div class="border-right" style="background-color: #f8f9fa" id="sidebar-wrapper">
-      <div class="sidebar-heading">
+    <div style="background-color: #ffffff; outline: 2px solid #c0c0c0" id="sidebar-wrapper">
+      <div style="padding: 0.44rem 1.25rem">
         <a href="index.html">  
           <img src="assets/logo.png" alt="Hex Logo" style="width:px;height:50px;display: block;margin-left: auto;margin-right: auto;">
         </a>
@@ -95,7 +95,11 @@
     <!-- Page Content -->
     <div id="page-content-wrapper">
 
-      <nav class="navbar navbar-expand-lg navbar-light bg-light border-bottom">
+      <!-- Header -->
+      
+      <!-- Note: Content of header copied from 'index.html',
+           managed by a python script -->
+      <nav class="navbar navbar-expand-lg navbar-light" style="border-bottom: 2px solid #c0c0c0;background-color: #ffffff">
         <button class="toggle-button" id="menu-toggle">Toggle Menu</button>
 
         <div class="home-location">
@@ -106,6 +110,7 @@
           </ul>
         </div>
       </nav>
+      <!-- /#header-wrapper -->
       
       <style>
         p {
