+++ index.html
@@ -95,6 +95,10 @@
     <!-- Page Content -->
     <div id="page-content-wrapper">
 
+      <!-- Header -->
+      
+      <!-- Note: Content of header copied from 'index.html',
+           managed by a python script -->
       <nav class="navbar navbar-expand-lg navbar-light" style="border-bottom: 2px solid #c0c0c0;background-color: #ffffff">
         <button class="toggle-button" id="menu-toggle">Toggle Menu</button>
 
@@ -106,6 +110,7 @@
           </ul>
         </div>
       </nav>
+      <!-- /#header-wrapper -->
 
       <div class="container-fluid" style="max-width: 1000px">
         <img src="assets/Profile%20Picture.jpg" alt="Profile Picture" style="height:200px; float: left; margin-right: 20px;border-radius: 20px">
@@ -119,8 +124,8 @@
         <p class="text-1" style="text-indent: 10px;margin-top: 10px;">This website serves as my online repository of finished work.</p>
         <p style="margin: 60px"></p>
         <h3 class="head-2">Professional Description</h3>
-        <p class="text-1">I'm a senior at Marquette University. In May of 2021 I plan to recieve a Bachelor of Science Degree in Electrical Engineering and complete my minor in Business Adminstration. I currently work as an undergraduate research assistant at Marquette University, developing a procedure for modifying an off-the-shelf incubator to make it research-grade. Throughout 2019, I worked at Extreme Engineering Solutions (X-ES) as a Design Verification Testing Co-op student, a position to which X-ES graciously allowed me to return the next summer after my plans to intern as an Analog Design Engineer at another company were cancelled to due COVID-19 concerns.</p>
-        <p class="text-1">X-ES developed my circuit board testing and analyzation skills, and gave me an understanding of how a single-board computer is designed and verified. I also have personal experience designing printed circuit boards using both Eagle and Atium in my own projects, which are documented here. Much of my work has circled around designing, verifying, assembling, and testing digital embedded solutions.</p>
+        <p class="text-1">I'm a project electrical engineer at Lutron, a lighting company specialized in lighting control. I work at Lutron's Boston location desiging and testing both analog and digital circuits in their products. The projects I've worked on have had moved rapidly (which keeps the work exciting) but also have focused on furthering Lutron's reputation for quality products (which makes the work rewarding).</p>
+        <p class="text-1">I graduated from Marquette University in May of 2021 with a Bachelor of Science Degree in Electrical Engineering and a minor in Business Administration. Both my professional and personal experiences in the latter half of my college career focused on design and testing of printed circuit boards for digital embedded solutions circuits. As part of Marquette's engineering program, I co-oped at Extreme Engineering Solutions in Madison, WI, where I performed design validation work on their single-board computers. Additionally, at Marquette, I worked as an undergraduate research assistant providing electrical support to eniromental experiments. Finally, during college and to a lesser degree after graduation I've gained experience working on my own engineering projects, which I have documented here.</p>
       </div>
     </div>
